orchestration/workflow-engine/sdlc_workflow.py
"""
SDLC Workflow Engine using LangGraph
Orchestrates the entire software development lifecycle through AI agents
"""

import operator
from typing import Annotated, List, Optional, TypedDict

from langchain_core.messages import BaseMessage
from langgraph.checkpoint import MemorySaver
from langgraph.graph import END, StateGraph

from ..models.workflow import StageResult


class SDLCState(TypedDict):
    """State maintained throughout the SDLC workflow"""

    project_idea: str
    current_stage: str
    stage_results: Annotated[List[StageResult], operator.add]
    messages: Annotated[List[BaseMessage], operator.add]
    error_count: int
    requires_escalation: bool
    final_artifacts: Optional[dict]


class SDLCWorkflowEngine:
    """Main workflow engine that orchestrates SDLC stages"""

    def __init__(self):
        self.checkpointer = MemorySaver()
        self.graph = None
        self._build_workflow()

    def _build_workflow(self):
        """Build the LangGraph workflow for SDLC stages"""
        workflow = StateGraph(SDLCState)

        # Define nodes for each SDLC stage
        workflow.add_node("market_validation", self.market_validation_node)
        workflow.add_node("solution_architecture", self.solution_architecture_node)
        workflow.add_node("product_management", self.product_management_node)
        workflow.add_node("software_architecture", self.software_architecture_node)
        workflow.add_node("junior_development", self.junior_development_node)
        workflow.add_node("senior_development", self.senior_development_node)
        workflow.add_node("quality_assurance", self.quality_assurance_node)
        workflow.add_node("deployment", self.deployment_node)

        # Add conditional edges for branching logic
        workflow.add_edge("market_validation", "solution_architecture")
        workflow.add_edge("solution_architecture", "product_management")
        workflow.add_edge("product_management", "software_architecture")
        workflow.add_edge("software_architecture", "junior_development")

        # Add conditional edge for escalation
        workflow.add_conditional_edges(
            "junior_development",
            self.check_escalation,
            {
                "escalate": "senior_development",
                "continue": "quality_assurance",
                "retry": "junior_development",
            },
        )

        workflow.add_edge("senior_development", "quality_assurance")
        workflow.add_edge("quality_assurance", "deployment")
        workflow.add_edge("deployment", END)

        # Set entry point
        workflow.set_entry_point("market_validation")

        # Compile the graph
        self.graph = workflow.compile(checkpointer=self.checkpointer)

    async def market_validation_node(self, state: SDLCState) -> SDLCState:
        """Market validation stage"""
        # TODO: Implement market validation agent logic
        state["current_stage"] = "market_validation"
        return state

    async def solution_architecture_node(self, state: SDLCState) -> SDLCState:
        """Solution architecture stage"""
        # TODO: Implement solution architecture agent logic
        state["current_stage"] = "solution_architecture"
        return state

    async def product_management_node(self, state: SDLCState) -> SDLCState:
        """Product management stage"""
        # TODO: Implement product management agent logic
        state["current_stage"] = "product_management"
        return state

    async def software_architecture_node(self, state: SDLCState) -> SDLCState:
        """Software architecture stage"""
        # TODO: Implement software architecture agent logic
        state["current_stage"] = "software_architecture"
        return state

    async def junior_development_node(self, state: SDLCState) -> SDLCState:
        """Junior developer implementation stage"""
        # TODO: Implement junior developer agent logic
        state["current_stage"] = "junior_development"
        return state

    async def senior_development_node(self, state: SDLCState) -> SDLCState:
        """Senior developer escalation stage"""
        # TODO: Implement senior developer agent logic
        state["current_stage"] = "senior_development"
        state["requires_escalation"] = False
        return state

    async def quality_assurance_node(self, state: SDLCState) -> SDLCState:
        """Quality assurance stage"""
        # TODO: Implement QA agent logic
        state["current_stage"] = "quality_assurance"
        return state

    async def deployment_node(self, state: SDLCState) -> SDLCState:
        """Deployment stage"""
        # TODO: Implement deployment agent logic
        state["current_stage"] = "deployment"
        return state

    def check_escalation(self, state: SDLCState) -> str:
        """Determine if escalation to senior developer is needed"""
        if state.get("requires_escalation", False):
            return "escalate"
        elif state.get("error_count", 0) > 3:
            return "retry"
        else:
            return "continue"

    async def run_workflow(self, project_idea: str, thread_id: str = "default"):
        """Execute the complete SDLC workflow"""
        initial_state = SDLCState(
            project_idea=project_idea,
            current_stage="start",
            stage_results=[],
            messages=[],
            error_count=0,
            requires_escalation=False,
            final_artifacts=None,
        )

        config = {"configurable": {"thread_id": thread_id}}

        async for event in self.graph.astream(initial_state, config):
            yield event
