requirements.txt
# Core dependencies for traditional pip users
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
crewai>=0.1.0
aider-chat>=0.20.0
gitpython>=3.1.40
pygithub>=2.1.1
celery[redis]>=5.3.4
redis>=5.0.1
pydantic>=2.5.0
python-telegram-bot>=20.7
discord-webhook>=1.3.0
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
websockets>=12.0
sqlalchemy>=2.0.23
alembic>=1.13.0
pgvector>=0.2.4
phoenix-arize>=2.0.0
agentops>=0.1.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.1
ruff>=0.1.9
python-dotenv>=1.0.0
